media: bcm2835-unicam: Support unpacking CSI format to 16bpp
BugLink: https://bugs.launchpad.net/bugs/1850876

The CSI packed formats are not the easiest to work with, and
the peripheral supports unpacking them to 16bpp (but NOT
shifting the data up into the MSBs).
Where V4L2 exposes a pixfmt for both packed and unpacked
formats advertise both as being supported, and unpack the
data in the peripheral.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.org>
(cherry picked from commit 299a673624cb1108c4f3b9dcdc833d769a434142
https://github.com/raspberrypi/linux.git rpi-5.3.y)
Signed-off-by: Hui Wang <hui.wang@canonical.com>
Signed-off-by: Khalid Elmously <khalid.elmously@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Sultan Alsawaf <sultan.alsawaf@canonical.com>
Signed-off-by: Khalid Elmously <khalid.elmously@canonical.com>